Worked on modularity and dynamic user input -- it is scuffed rn
