+++ server/controllers/util/metricDataQueries.js
@@ -1,60 +1,14 @@
-// { functionName, stat, sortBy, period, startDate } = req.body;
-
 let idCount = 0;
-//needs to be filtered after receiving data from front end: funciton name, metric name,time period (seconds), date range, scan by (sort by oldest to newest or newest to oldest)
+const queryCreator = {}
 
-const formattedStartDate = dateConverter(startDate); // pass in start date from frontend to format into usable date
-
-//START TIME AND END TIME
-const dateConverter = (date) => {
-  if(date[1] === 'w'){
-    const startDate = Date();
-    startDate.setDate(startDate.getDate() - 7);
-    return startDate;
-  } else if(date[1] === 'h'){
-    const fraction = Number(date[0]) / 24
-    const startDate = Date();
-    startDate.setDate(startDate.getDate() - fraction);
-    return startDate;
-  } else{
-    const startDate = Date();
-    startDate.setDate(startDate.getDate() - Number(date[0]));
-    return startDate;
-  }
-}
-
-
-let period; //delete this once req.body is functional
-let finalPeriod;
-const timePeriodConverter = (period) => {
-  if(period === '1 second' || period === '5 seconds' || period === '10 seconds' || period === '30 seconds') {
-    if(period.length === 8) {
-      finalPeriod = Number(period[0]);
-    } else {
-      finalPeriod = Number(period[0]) + Number(period[1]);
-    }
-  } 
-  if(period === '1 minute' || period === '5 minutes' || period === '15 minutes') {
-    if(period.length === 8) {
-      finalPeriod = Number(period[0]);
-    } else {
-      finalPeriod = Number(period[0]) + Number(period[1]);
-    }
-  };
-  return finalPeriod;
-};
-
-const queryCreator = (functionName, stat, sortBy, period, startDate) => {
-
-}
-
+queryCreator.createQuery = (funcName, stat, sortBy, period, startDate) => {
 const metricDataQueries = {};
 
 //METRICS
 metricDataQueries.duration = {
   // GetMetricDataInput
   MetricDataQueries: [
-    // MetricDataQueries // required
+  // MetricDataQueries // required
     {
       Id: `duration${idCount}`, // required
       MetricStat: {
@@ -64,19 +18,19 @@ metricDataQueries.duration = {
           Dimensions: [
             {
               Name: `FunctionName`, // required  - can be hardcoded as funciton name
-              Value: 'secondFunction', // required  //function name - needs to be actual function name from front end. req.queried
+              Value: `${funcName}`, // required  //function name - needs to be actual function name from front end. req.queried
             },
           ],
         },
-        Period: 900, // required (also req.body)
-        Stat: 'Average', // required (Average, min/max, sum, sample count, etc) (also req.body)
+        Period: period, // required (also req.body)
+        Stat: `${stat}`, // required (Average, min/max, sum, sample count, etc) (also req.body)
       },
       Label: 'Lambda Duration secondFunction', //function and MetricName need to be dynamic
     },
   ],
-  StartTime: sevenDaysAgo, // `${formattedStartDate}`
+  StartTime: `${startDate}`,
   EndTime: new Date(), // required (will always be now)
-  ScanBy: 'TimestampDescending' || 'TimestampAscending',
+  ScanBy: `${sortBy}`,
   LabelOptions: {
     // LabelOptions
     Timezone: '-0400', //comes from region
@@ -96,19 +50,19 @@ metricDataQueries.invocations = {
           Dimensions: [
             {
               Name: `FunctionName`, // required  - can be hardcoded as funciton name
-              Value: 'secondFunction', // required  //function name - needs to be actual function name from front end. req.queried
+              Value: `${funcName}`, // required  //function name - needs to be actual function name from front end. req.queried
             },
           ],
         },
-        Period: 900, // required (also req.body)
-        Stat: 'Average', // required (Average, min/max, sum, sample count, etc) (also req.body)
+        Period: period, // required (also req.body)
+        Stat: `${stat}`, // required (Average, min/max, sum, sample count, etc) (also req.body)
       },
       Label: 'Lambda Invocations secondFunction', //function and MetricName need to be dynamic
     },
   ],
-  StartTime: sevenDaysAgo, // `${formattedStartDate}`
+  StartTime: `${startDate}`,
   EndTime: new Date(), // required (will always be now)
-  ScanBy: 'TimestampDescending' || 'TimestampAscending',
+  ScanBy: `${sortBy}`,
   LabelOptions: {
     // LabelOptions
     Timezone: '-0400', //comes from region
@@ -128,19 +82,19 @@ metricDataQueries.throttles = {
           Dimensions: [
             {
               Name: `FunctionName`, // required  - can be hardcoded as funciton name
-              Value: 'secondFunction', // required  //function name - needs to be actual function name from front end. req.queried
+              Value: `${funcName}`, // required  //function name - needs to be actual function name from front end. req.queried
             },
           ],
         },
-        Period: 900, // required (also req.body)
-        Stat: 'Average', // required (Average, min/max, sum, sample count, etc) (also req.body)
+        Period: period, // required (also req.body)
+        Stat: `${stat}`, // required (Average, min/max, sum, sample count, etc) (also req.body)
       },
       Label: 'Lambda Throttles secondFunction', //function and MetricName need to be dynamic
     },
   ],
-  StartTime: sevenDaysAgo, // `${formattedStartDate}`
+  StartTime: `${startDate}`,
   EndTime: new Date(), // required (will always be now)
-  ScanBy: 'TimestampDescending' || 'TimestampAscending',
+  ScanBy: `${sortBy}`,
   LabelOptions: {
     // LabelOptions
     Timezone: '-0400', //comes from region
@@ -160,19 +114,19 @@ metricDataQueries.errors = {
           Dimensions: [
             {
               Name: `FunctionName`, // required  - can be hardcoded as funciton name
-              Value: 'secondFunction', // required  //function name - needs to be actual function name from front end. req.queried
+              Value: `${funcName}`, // required  //function name - needs to be actual function name from front end. req.queried
             },
           ],
         },
-        Period: 900, // required (also req.body)
-        Stat: 'Average', // required (Average, min/max, sum, sample count, etc) (also req.body)
+        Period: period, // required (also req.body)
+        Stat: `${stat}`, // required (Average, min/max, sum, sample count, etc) (also req.body)
       },
       Label: 'Lambda Errors secondFunction', //function and MetricName need to be dynamic
     },
   ],
-  StartTime: sevenDaysAgo, // `${formattedStartDate}`
+  StartTime: `${startDate}`,
   EndTime: new Date(), // required (will always be now)
-  ScanBy: 'TimestampDescending' || 'TimestampAscending',
+  ScanBy: `${sortBy}`,
   LabelOptions: {
     // LabelOptions
     Timezone: '-0400', //comes from region
@@ -192,17 +146,17 @@ metricDataQueries.concurrentExecutions = {
           Dimensions: [
             {
               Name: `FunctionName`, // required  - can be hardcoded as funciton name
-              Value: 'secondFunction', // required  //function name - needs to be actual function name from front end. req.queried
+              Value: `${funcName}`, // required  //function name - needs to be actual function name from front end. req.queried
             },
           ],
         },
-        Period: 900, // required (also req.body)
-        Stat: 'Average', // required (Average, min/max, sum, sample count, etc) (also req.body)
+        Period: period, // required (also req.body)
+        Stat: `${stat}`, // required (Average, min/max, sum, sample count, etc) (also req.body)
       },
       Label: 'Lambda Concurrent Executions secondFunction', //function and MetricName need to be dynamic
     },
   ],
-  StartTime: sevenDaysAgo, // `${formattedStartDate}`
+  StartTime: startDate, // `${formattedStartDate}`
   EndTime: new Date(), // required (will always be now)
   ScanBy: 'TimestampDescending' || 'TimestampAscending',
   LabelOptions: {
@@ -213,4 +167,6 @@ metricDataQueries.concurrentExecutions = {
 
 idCount += 1;
 
-module.exports = metricDataQueries;
+}
+
+module.exports = queryCreator;
